service-api/web/serializers.py
from django.db.models import fields
from rest_framework import serializers
from django.core.exceptions import PermissionDenied, ValidationError
from django.utils.translation import gettext_lazy as _
from rest_framework import status
from web.models import *
from web.models.option import *
from rest_framework.validators import UniqueTogetherValidator
