service-api/requirements.txt
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
b2sdk==1.0.2
certifi==2022.6.15
chardet==3.0.4
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
Deprecated==1.2.13
Django==3.2.3
django-admin-rangefilter==0.6.3
django-b2==0.7.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-mongolog==0.9.4
django-oauth-toolkit==1.2.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-rq==2.3.2
djangorestframework==3.12.4
drf-yasg==1.21.5
et-xmlfile==1.1.0
gunicorn==20.0.4
idna==2.6
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
logfury==1.0.1
MarkupSafe==2.1.2
oauthlib==3.2.0
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.1.0
pymemcache==3.4.4
pymongo==3.11.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.17.1
python-gettext==4.0
pytz==2022.2.1
redis==4.3.4
requests==2.18.4
rq==1.11.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.22
wrapt==1.14.1
zipp==3.8.1
